requirements.txt
Django
Celery
Redis
django-crispy-forms
crispy-bootstrap4
psycopg2-binary
aiohttp
Pillow
djangorestframework